Allow toggling of derived fields
